+++ v3/lints/rfc/lint_dnsname_hyphen_in_sld_test.go
@@ -1,4 +1,4 @@
-package cabf_br
+package rfc
 
 /*
  * ZLint Copyright 2021 Regents of the University of Michigan
