+++ v3/lints/rfc/lint_dnsname_label_too_long_test.go
@@ -1,4 +1,4 @@
-package cabf_br
+package rfc
 
 /*
  * ZLint Copyright 2021 Regents of the University of Michigan
