+++ v3/lints/rfc/lint_dnsname_underscore_in_trd.go
@@ -1,4 +1,4 @@
-package cabf_br
+package rfc
 
 /*
  * ZLint Copyright 2021 Regents of the University of Michigan
@@ -22,46 +22,37 @@ import (
 	"github.com/zmap/zlint/v3/util"
 )
 
-type DNSNameUnderscoreInSLD struct{}
+type DNSNameUnderscoreInTRD struct{}
 
 func init() {
 	lint.RegisterLint(&lint.Lint{
-		Name:          "e_dnsname_underscore_in_sld",
+		Name:          "w_dnsname_underscore_in_trd",
 		Description:   "DNSName MUST NOT contain underscore characters",
-		Citation:      "BRs: 7.1.4.2.1",
-		Source:        lint.CABFBaselineRequirements,
+		Citation:      "RFC5280: 4.1.2.6",
+		Source:        lint.RFC5280,
 		EffectiveDate: util.RFC5280Date,
-		Lint:          &DNSNameUnderscoreInSLD{},
+		Lint:          &DNSNameUnderscoreInTRD{},
 	})
 }
 
-func (l *DNSNameUnderscoreInSLD) Initialize() error {
+func (l *DNSNameUnderscoreInTRD) Initialize() error {
 	return nil
 }
 
-func (l *DNSNameUnderscoreInSLD) CheckApplies(c *x509.Certificate) bool {
+func (l *DNSNameUnderscoreInTRD) CheckApplies(c *x509.Certificate) bool {
 	return util.IsSubscriberCert(c) && util.DNSNamesExist(c)
 }
 
-func (l *DNSNameUnderscoreInSLD) Execute(c *x509.Certificate) *lint.LintResult {
-	if c.Subject.CommonName != "" && !util.CommonNameIsIP(c) {
-		domainInfo := c.GetParsedSubjectCommonName(false)
-		if domainInfo.ParseError != nil {
-			return &lint.LintResult{Status: lint.NA}
-		}
-		if strings.Contains(domainInfo.ParsedDomain.SLD, "_") {
-			return &lint.LintResult{Status: lint.Error}
-		}
-	}
-
+func (l *DNSNameUnderscoreInTRD) Execute(c *x509.Certificate) *lint.LintResult {
 	parsedSANDNSNames := c.GetParsedDNSNames(false)
 	for i := range c.GetParsedDNSNames(false) {
 		if parsedSANDNSNames[i].ParseError != nil {
 			return &lint.LintResult{Status: lint.NA}
 		}
-		if strings.Contains(parsedSANDNSNames[i].ParsedDomain.SLD, "_") {
-			return &lint.LintResult{Status: lint.Error}
+		if strings.Contains(parsedSANDNSNames[i].ParsedDomain.TRD, "_") {
+			return &lint.LintResult{Status: lint.Warn}
 		}
 	}
+
 	return &lint.LintResult{Status: lint.Pass}
 }
